+++ src/models/Guest.php
@@ -6,7 +6,7 @@ use craft\base\Model;
 
 class Guest extends Model
 {
-    public $name, $email, $phone, $notes;
+    public $name, $email, $phone, $notes, $redirect;
     public $state = 'waiting';
 
     public function init()
@@ -17,7 +17,7 @@ class Guest extends Model
     public function rules()
     {
         return [
-            [['name', 'state'], 'required'],
+            [['name', 'state', 'redirect'], 'required'],
         ];
     }
 
@@ -72,6 +72,17 @@ class Guest extends Model
     public function setState($value)
     {
         $this->state = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param $value
+     * @return $this
+     */
+    public function setRedirect($value)
+    {
+        $this->redirect = $value;
 
         return $this;
     }